log_api: merge log_api into oko_api
