--- backend/log_api/jest.config.mjs
@@ -1,26 +1,0 @@
-import { createDefaultEsmPreset } from "ts-jest";
-
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-export default {
-  ...createDefaultEsmPreset(),
-  testEnvironment: "node",
-  modulePathIgnorePatterns: ["<rootDir>/dist/"],
-  moduleNameMapper: {
-    "^@oko-wallet-log-api/(.*)$": "<rootDir>/src/$1",
-  },
-  extensionsToTreatAsEsm: [".ts"],
-  globals: {
-    "ts-jest": {
-      useESM: true,
-    },
-  },
-  transform: {
-    "^.+\\.tsx?$": [
-      "ts-jest",
-      {
-        useESM: true,
-      },
-    ],
-  },
-  transformIgnorePatterns: ["node_modules/(?!(@oko-wallet/.*)/)"],
-};
