--- backend/log_api/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "@oko-wallet/log-api",
-  "private": true,
-  "main": "./src/index.ts",
-  "type": "module",
-  "scripts": {
-    "typecheck": "tsc --noEmit",
-    "jest": "NODE_OPTIONS=--experimental-vm-modules jest"
-  },
-  "dependencies": {
-    "@oko-wallet/bytes": "^0.1.2-alpha.1",
-    "@oko-wallet/oko-api-error-codes": "workspace:*",
-    "@oko-wallet/oko-api-server-state": "workspace:*",
-    "@oko-wallet/oko-pg-interface": "workspace:*",
-    "@oko-wallet/oko-types": "^0.1.2-alpha.1",
-    "@oko-wallet/postgres-lib": "workspace:*",
-    "@oko-wallet/stdlib-js": "0.0.2-rc.41",
-    "cors": "^2.8.5",
-    "dayjs": "^1.11.18",
-    "dotenv": "^16.4.5",
-    "express": "^4.21.1",
-    "express-rate-limit": "8.2.1",
-    "helmet": "^8.0.0",
-    "jsonwebtoken": "^9.0.2",
-    "morgan": "^1.10.0",
-    "pg": "^8.16.3",
-    "swagger-ui-express": "^5.0.1",
-    "uuid": "^13.0.0",
-    "winston": "^3.17.0",
-    "winston-elasticsearch": "^0.19.0",
-    "zod": "^4.1.12"
-  },
-  "devDependencies": {
-    "@types/cors": "^2.8.17",
-    "@types/express": "^5.0.6",
-    "@types/express-rate-limit": "^6.0.2",
-    "@types/jest": "^29.5.14",
-    "@types/jsonwebtoken": "^9.0.7",
-    "@types/morgan": "^1.9.9",
-    "@types/node": "^24.10.1",
-    "@types/supertest": "^6.0.3",
-    "@types/swagger-jsdoc": "^6.0.4",
-    "@types/swagger-ui-express": "^4.1.8",
-    "jest": "^30.1.3",
-    "supertest": "^7.0.0",
-    "ts-jest": "^29.4.5",
-    "typescript": "^5.8.3"
-  }
-}
