--- backend/log_api/src/database/test_config.ts
@@ -1,10 +1,0 @@
-import type { PgDatabaseConfig } from "@oko-wallet/postgres-lib";
-
-export const testPgConfig: PgDatabaseConfig = {
-  database: "oko_dev",
-  host: "localhost",
-  password: "postgres",
-  user: "postgres",
-  port: 5432,
-  ssl: false,
-};
