--- backend/log_api/src/global.d.ts
@@ -1,3 +1,0 @@
-/// <reference types="@oko-wallet/oko-api-server-state/src/global" />
-
-export {};
