--- backend/log_api/src/index.ts
@@ -1,1 +1,0 @@
-export * from "./routes";
