+++ backend/oko_api/server/src/routes/index.ts
@@ -2,13 +2,13 @@ import { type Express } from "express";
 import { makeCustomerRouter } from "@oko-wallet/ct-dashboard-api";
 import { makeOkoAdminRouter } from "@oko-wallet/admin-api";
 import { makeUserRouter } from "@oko-wallet/user-dashboard-api";
-import { makeLogRouter } from "@oko-wallet/log-api";
 import { getStatus } from "./get_status";
 import { makeSocialLoginRouter } from "./social_login_v1";
 import { makeSocialLoginV2Router } from "./social_login_v2";
 import { makeTSSRouterV1 } from "./tss_v1";
 import { makeTSSRouterV2 } from "./tss_v2";
 import { makeAttachedV1Router } from "./attached_v1";
+import { makeLogRouterV1 } from "./log_v1";
 
 export function setRoutes(app: Express) {
   app.use("/customer_dashboard/v1", makeCustomerRouter());
@@ -23,15 +23,7 @@ export function setRoutes(app: Express) {
 
   app.use("/user_dashboard/v1", makeUserRouter());
 
-  app.use(
-    "/log/v1",
-    makeLogRouter({
-      esUrl: app.locals.es_url,
-      esIndex: app.locals.es_client_index,
-      esUsername: app.locals.es_username,
-      esPassword: app.locals.es_password,
-    }),
-  );
+  app.use("/log/v1", makeLogRouterV1());
 
   app.use("/social-login/v1", makeSocialLoginRouter());
