+++ backend/oko_api/server/src/testing/app.ts
@@ -6,11 +6,22 @@ import cors from "cors";
 import { makeTSSRouterV1 } from "@oko-wallet-api/routes/tss_v1";
 import { makeTSSRouterV2 } from "@oko-wallet-api/routes/tss_v2";
 
-export function makeApp(env: {
+export interface TestEnvsTss {
   JWT_SECRET: string;
   JWT_EXPIRES_IN: string;
   ENCRYPTION_SECRET: string;
-}) {
+}
+
+export interface TestEnvsLog {
+  ES_URL: string;
+  ES_INDEX: string;
+  ES_USERNAME: string;
+  ES_PASSWORD: string;
+}
+
+export type TestEnvs = TestEnvsTss | TestEnvsLog;
+
+export function makeApp(env: TestEnvs) {
   const app = express();
 
   app.use(morgan("dev"));
@@ -25,9 +36,10 @@ export function makeApp(env: {
   app.use("/tss/v1", makeTSSRouterV1);
   app.use("/tss/v2", makeTSSRouterV2);
 
-  app.locals.jwt_secret = env.JWT_SECRET;
-  app.locals.jwt_expires_in = env.JWT_EXPIRES_IN;
-  app.locals.encryption_secret = env.ENCRYPTION_SECRET;
+  const e = env as any;
+  app.locals.jwt_secret = e.JWT_SECRET;
+  app.locals.jwt_expires_in = e.JWT_EXPIRES_IN;
+  app.locals.encryption_secret = e.ENCRYPTION_SECRET;
 
   return app;
 }
