+++ lerna.json
@@ -10,7 +10,6 @@
     "backend/oko_api/error_codes",
     "backend/oko_api_server_state",
     "backend/oko_pg_interface",
-    "backend/log_api",
     "backend/ct_dashboard_api",
     "backend/user_dashboard_api",
     "common/oko_types",
