+++ package.json
@@ -30,7 +30,6 @@
       "backend/oko_api_error_codes",
       "backend/oko_api/server",
       "backend/oko_api_server_state",
-      "backend/log_api",
       "backend/ct_dashboard_api",
       "backend/oko_pg_interface",
       "backend/admin_api",
